+++ src/visual.ts
@@ -149,7 +149,8 @@ module powerbi.extensibility.visual {
             StartXpoint: 10,
             StartYpoint: 20,
             ElementWidth: 30,
-            MinCellWidth: 30,
+            MinCellWidth: 45,
+            MinCellHeight: 30,
             MaxCellHeight: 60,
             PeriodSlicerRectWidth: 15,
             PeriodSlicerRectHeight: 23,
@@ -232,6 +233,7 @@ module powerbi.extensibility.visual {
         private datePeriod: ITimelineDatePeriod;
         private prevFilteredStartDate: Date | null = null;
         private prevFilteredEndDate: Date | null = null;
+        private prevGranularity: GranularityType | null = null;
 
         private initialized: boolean;
 
@@ -268,7 +270,7 @@ module powerbi.extensibility.visual {
             this.timelineProperties = {
                 textYPosition: Timeline.DefaultTextYPosition,
                 cellsYPosition: Timeline.TimelineMargins.TopMargin
-                * Timeline.CellsYPositionFactor + Timeline.CellsYPositionOffset,
+                    * Timeline.CellsYPositionFactor + Timeline.CellsYPositionOffset,
                 topMargin: Timeline.TimelineMargins.TopMargin,
                 bottomMargin: Timeline.TimelineMargins.BottomMargin,
                 leftMargin: Timeline.TimelineMargins.LeftMargin,
@@ -337,31 +339,6 @@ module powerbi.extensibility.visual {
                 this.selectionManager.clear();
 
                 if (this.timelineData) {
-                    this.timelineData.selectionStartIndex = Timeline.DefaultSelectionStartIndex;
-
-                    this.timelineData.selectionEndIndex =
-                        this.timelineData.currentGranularity.getDatePeriods().length - 1;
-
-                    const hasIrregularDataPoint: boolean = this.timelineData.timelineDatapoints
-                        .some((dataPoint: TimelineDatapoint) => dataPoint.index % 1 !== 0);
-
-                    if (hasIrregularDataPoint) {
-                        this.selectPeriod(this.timelineData.currentGranularity.getType());
-                    }
-                    else {
-                        Timeline.updateCursors(this.timelineData, this.timelineProperties.cellWidth);
-
-                        this.fillCells(this.settings.cells);
-
-                        this.renderCursors(
-                            this.timelineData,
-                            this.timelineProperties.cellHeight,
-                            this.timelineProperties.cellsYPosition);
-
-                        this.renderTimeRangeText(this.timelineData, this.settings.rangeHeader);
-                        this.fillColorGranularity(this.settings.granularity);
-                    }
-
                     this.clearSelection(this.timelineData.filterColumnTarget);
                 }
             }
@@ -543,7 +520,8 @@ module powerbi.extensibility.visual {
             viewport: IViewport,
             timelineProperties: TimelineProperties,
             timelineMargins: TimelineMargins,
-            scaleSizeAdjustment: ScaleSizeAdjustment): void {
+            scaleSizeAdjustment: ScaleSizeAdjustment,
+            labelFontSize: number = 9): void {
 
             timelineProperties.cellsYPosition = timelineProperties.textYPosition;
 
@@ -568,7 +546,7 @@ module powerbi.extensibility.visual {
 
             if (scaleSizeAdjustment.show) {
                 height = Math.max(
-                    timelineMargins.MinCellWidth,
+                    timelineMargins.MinCellHeight,
                     Math.min(
                         timelineMargins.MaxCellHeight,
                         maxHeight,
@@ -576,7 +554,7 @@ module powerbi.extensibility.visual {
                         - timelineProperties.cellsYPosition
                         - Timeline.TimelinePropertiesHeightOffset));
             } else {
-                height = timelineMargins.MinCellWidth;
+                height = timelineMargins.MinCellHeight;
             }
 
             width = Math.max(
@@ -592,29 +570,8 @@ module powerbi.extensibility.visual {
         }
 
         private createTimelineData(dataView: DataView) {
-            let startDate: Date,
-                endDate: Date,
-                datePeriod: TimelineDatePeriodBase,
-                resetedStartDate: Date,
-                resetedEndDate: Date;
-
-            datePeriod = this.settings.general.datePeriod as TimelineDatePeriodBase;
-
-            if (datePeriod.startDate && datePeriod.endDate) {
-                resetedStartDate = Utils.resetTime(datePeriod.startDate);
-                resetedEndDate = Utils.resetTime(Utils.getEndOfThePreviousDate(datePeriod.endDate));
-
-                startDate = this.datePeriod.startDate < resetedStartDate
-                    ? this.datePeriod.startDate
-                    : resetedStartDate;
-
-                endDate = this.datePeriod.endDate > resetedEndDate
-                    ? this.datePeriod.endDate
-                    : resetedEndDate;
-            } else {
-                startDate = this.datePeriod.startDate;
-                endDate = this.datePeriod.endDate;
-            }
+            let startDate = this.datePeriod.startDate;
+            let endDate = this.datePeriod.endDate;
 
             if (!this.initialized) {
                 this.drawGranular(this.timelineProperties, this.settings.granularity.granularity);
@@ -667,44 +624,57 @@ module powerbi.extensibility.visual {
         public static selectCurrentPeriod(
             datePeriod: ITimelineDatePeriod,
             granularity: GranularityType,
-            calendar) {
-            let currentDate: Date = Utils.resetTime(new Date());
-            let startDate: Date = currentDate,
-                endDate: Date;
+            calendar
+        ) {
+            return this.selectPeriod(datePeriod, granularity, calendar, Utils.resetTime(new Date()));
+        }
 
+        public static selectPeriod(
+            datePeriod: ITimelineDatePeriod,
+            granularity: GranularityType,
+            calendar,
+            periodDate: Date
+        ) {
+            let startDate: Date = periodDate,
+                endDate: Date;
             switch (granularity) {
                 case GranularityType.day:
-                    endDate = calendar.getNextDate(currentDate);
+                    endDate = calendar.getNextDate(periodDate);
                     break;
                 case GranularityType.week:
-                    ({startDate, endDate} = calendar.getWeekPeriod(currentDate));
+                    ({ startDate, endDate } = calendar.getWeekPeriod(periodDate));
                     break;
                 case GranularityType.month:
-                    ({startDate, endDate} = calendar.getMonthPeriod(currentDate));
+                    ({ startDate, endDate } = calendar.getMonthPeriod(periodDate));
                     break;
                 case GranularityType.quarter:
-                    ({startDate, endDate} = calendar.getQuarterPeriod(currentDate));
+                    ({ startDate, endDate } = calendar.getQuarterPeriod(periodDate));
                     break;
                 case GranularityType.year:
-                    ({startDate, endDate} = calendar.getYearPeriod(currentDate));
+                    ({ startDate, endDate } = calendar.getYearPeriod(periodDate));
                     break;
             }
 
-            const checkDatesForNoneDayGranularity: boolean =
-                datePeriod.startDate <= startDate || startDate <= datePeriod.endDate ||
-                datePeriod.startDate <= endDate || endDate <= datePeriod.endDate;
+            if (granularity === GranularityType.day) {
+                const checkDatesForDayGranularity: boolean =
+                    datePeriod.startDate <= startDate && endDate <= datePeriod.endDate ||
+                    startDate.toString() === datePeriod.endDate.toString();
 
-            const checkDatesForDayGranularity: boolean =
-                datePeriod.startDate <= startDate && endDate <= datePeriod.endDate ||
-                startDate.toString() === datePeriod.endDate.toString();
+                if (!checkDatesForDayGranularity) {
+                    startDate = null;
+                    endDate = null;
+                }
+            } else {
+                const endDateAvailable = (datePeriod.startDate <= startDate && startDate <= datePeriod.endDate);
+                const startDateAvailable = (datePeriod.startDate <= endDate && endDate <= datePeriod.endDate);
 
-            if (!(checkDatesForNoneDayGranularity && granularity !== GranularityType.day ||
-                checkDatesForDayGranularity && granularity === GranularityType.day)) {
-                startDate = null;
-                endDate = null;
+                if (!startDateAvailable && !endDateAvailable) {
+                    startDate = null;
+                    endDate = null;
+                }
             }
 
-            return {startDate, endDate};
+            return { startDate, endDate };
         }
 
         public static areVisualUpdateOptionsValid(options: VisualUpdateOptions): boolean {
@@ -749,6 +719,7 @@ module powerbi.extensibility.visual {
                 this.clearData();
                 return;
             }
+
             this.options = options;
             this.dataView = options.dataViews[0];
 
@@ -788,18 +759,27 @@ module powerbi.extensibility.visual {
             const latestAvailableDate: boolean = this.settings.forceSelection.latestAvailableDate;
             const isUserSelection: boolean = this.settings.general.isUserSelection;
             const target: IFilterColumnTarget = this.timelineData.filterColumnTarget;
-
             if (!isUserSelection) {
                 if (currentForceSelection) {
                     ({
                         startDate: filterDatePeriod.startDate,
                         endDate: filterDatePeriod.endDate
                     } = Timeline.selectCurrentPeriod(datePeriod, granularity, this.calendar));
-                }
-            } else {
-                if (latestAvailableDate) {
+                } else if (latestAvailableDate) {
                     filterDatePeriod.endDate = adaptedDataEndDate;
                 }
+
+                if (currentForceSelection && latestAvailableDate && (!filterDatePeriod.startDate || !filterDatePeriod.endDate)) {
+                    ({
+                        startDate: filterDatePeriod.startDate,
+                        endDate: filterDatePeriod.endDate
+                    } = Timeline.selectPeriod(datePeriod, granularity, this.calendar, this.datePeriod.endDate));
+                }
+            } else if (this.prevGranularity !== granularity && currentForceSelection) {
+                ({
+                    startDate: filterDatePeriod.startDate,
+                    endDate: filterDatePeriod.endDate
+                } = Timeline.selectCurrentPeriod(datePeriod, granularity, this.calendar));
             }
 
             const filterWasChanged: boolean =
@@ -809,11 +789,16 @@ module powerbi.extensibility.visual {
             if ((!isUserSelection && filterWasChanged) ||
                 (isUserSelection && filterWasChanged && latestAvailableDate) ||
                 (!this.initialized && !currentForceSelection)) {
-                this.applyDatePeriod(filterDatePeriod.startDate, filterDatePeriod.endDate, target, isUserSelection);
+                 this.applyDatePeriod(filterDatePeriod.startDate, filterDatePeriod.endDate, target, isUserSelection);
+            }
+
+            if (this.prevGranularity && this.prevGranularity !== granularity && latestAvailableDate) {
+                filterDatePeriod.endDate = adaptedDataEndDate;
             }
 
             this.prevFilteredStartDate = filterDatePeriod.startDate;
             this.prevFilteredEndDate = filterDatePeriod.endDate;
+            this.prevGranularity = granularity;
 
             if (!this.initialized) {
                 this.initialized = true;
@@ -835,7 +820,7 @@ module powerbi.extensibility.visual {
                 options);
         }
 
-        private selectPeriod(granularityType: GranularityType): void {
+        public selectPeriod(granularityType: GranularityType): void {
             if (this.timelineData.currentGranularity.getType() !== granularityType) {
                 this.host.persistProperties({
                     merge: [{
@@ -846,6 +831,7 @@ module powerbi.extensibility.visual {
                 });
 
                 this.settings.granularity.granularity = granularityType;
+                return;
             }
 
             this.redrawPeriod(granularityType);
@@ -868,7 +854,8 @@ module powerbi.extensibility.visual {
                 this.initialized,
                 timelineFormat,
                 this.options.viewport,
-                this.calendar);
+                this.calendar,
+                this.settings);
         }
 
         private static isDataViewValid(dataView): boolean {
@@ -897,7 +884,8 @@ module powerbi.extensibility.visual {
             initialized: boolean,
             timelineSettings: VisualSettings,
             viewport: IViewport,
-            previousCalendar: Calendar): Calendar {
+            previousCalendar: Calendar,
+            setting: VisualSettings): Calendar {
 
             if (this.isDataViewValid(dataView)) {
                 return null;
@@ -963,6 +951,7 @@ module powerbi.extensibility.visual {
             }
 
             if (isCalendarChanged && startDate && endDate) {
+                Utils.unseparateSelection(timelineData.currentGranularity.getDatePeriods());
                 Utils.separateSelection(timelineData, startDate, endDate);
             }
 
@@ -993,7 +982,8 @@ module powerbi.extensibility.visual {
                 viewport,
                 timelineProperties,
                 Timeline.TimelineMargins,
-                timelineSettings.scaleSizeAdjustment
+                timelineSettings.scaleSizeAdjustment,
+                setting.labels.textSize
             );
 
             Timeline.updateCursors(timelineData, timelineProperties.cellWidth);
@@ -1082,7 +1072,6 @@ module powerbi.extensibility.visual {
                     this.yearLabelsSelection,
                     yPos,
                     granularityType === 0);
-
                 yPos += yDiff;
             }
 
@@ -1092,7 +1081,6 @@ module powerbi.extensibility.visual {
                     this.quarterLabelsSelection,
                     yPos,
                     granularityType === 1);
-
                 yPos += yDiff;
             }
 
@@ -1102,7 +1090,6 @@ module powerbi.extensibility.visual {
                     this.monthLabelsSelection,
                     yPos,
                     granularityType === 2);
-
                 yPos += yDiff;
             }
 
@@ -1199,7 +1186,7 @@ module powerbi.extensibility.visual {
                         return (label.id + Timeline.LabelIdOffset) * this.timelineProperties.cellWidth;
                     },
                     y: this.timelineProperties.textYPosition
-                    + (1 + index) * fromPointToPixel(this.settings.labels.textSize),
+                        + (1 + index) * fromPointToPixel(this.settings.labels.textSize),
                     fill: this.settings.labels.fontColor
                 })
                 .append("title")
@@ -1538,7 +1525,7 @@ module powerbi.extensibility.visual {
                     .classed(Timeline.TimelineSelectors.SelectionRangeContainer.className, true)
                     .attr({
                         x: GranularityNames.length
-                        * (this.timelineProperties.elementWidth + this.timelineProperties.leftMargin),
+                            * (this.timelineProperties.elementWidth + this.timelineProperties.leftMargin),
                         y: Timeline.DefaultRangeTextSelectionY,
                         fill: rangeHeaderSettings.fontColor
                     })